requirements.txt
click
PYyaml